+++ bioimageio/spec/_description.py
@@ -1,3 +1,4 @@
+from copy import deepcopy
 from pathlib import Path
 from types import MappingProxyType
 from typing import Any, Literal, Optional, TypeVar, Union
@@ -5,11 +6,15 @@ from typing import Any, Literal, Optional, TypeVar, Union
 from pydantic import Discriminator
 from typing_extensions import Annotated
 
+from bioimageio.spec._internal.io_utils import open_bioimageio_yaml
+from bioimageio.spec._internal.validation_context import ValidationContext
+
 from ._description_impl import DISCOVER, build_description_impl, get_rd_class_impl
 from ._internal.common_nodes import InvalidDescr
 from ._internal.io import BioimageioYamlContent, BioimageioYamlSource
+from ._internal.io_utils import write_yaml
 from ._internal.types import FormatVersionPlaceholder
-from ._internal.validation_context import ValidationContext, validation_context_var
+from ._internal.validation_context import validation_context_var
 from .application import (
     AnyApplicationDescr,
     ApplicationDescr,
@@ -71,6 +76,16 @@ def dump_description(
 RD = TypeVar("RD", bound=ResourceDescr)
 
 
+LATEST_DESCRIPTIONS_MAP = MappingProxyType(
+    {
+        None: GenericDescr,
+        "generic": GenericDescr,
+        "application": ApplicationDescr,
+        "dataset": DatasetDescr,
+        "notebook": NotebookDescr,
+        "model": ModelDescr,
+    }
+)
 DESCRIPTIONS_MAP = MappingProxyType(
     {
         None: MappingProxyType(
@@ -190,10 +205,31 @@ def update_format(
     source: BioimageioYamlSource,
     *,
     output_path: Optional[Path] = None,
-    target_format_version: Union[Literal["latest"], str] = LATEST,
+    # target_format_version: Union[Literal["latest"], str] = LATEST,
+    # TODO: support updating to non-latest format versions
 ) -> BioimageioYamlContent:
-    """update a bioimageio.yaml file without validating it"""
-    raise NotImplementedError("Oh no! This feature is not yet implemented")
+    """Update of an rdf.yaml (or bioimageio.yaml) file without validating it.
+
+    Note: This function does not update patch version related changes.
+        Use `load_description` to update the patch version.
+    """
+
+    if isinstance(source, dict):
+        content = deepcopy(source)
+    else:
+        opened = open_bioimageio_yaml(source)
+        content = opened.content
+
+    source_type = s if isinstance((s := content.get("type")), str) else "unknown"
+    rd_class = LATEST_DESCRIPTIONS_MAP.get(source_type, GenericDescr)
+    converted_content = rd_class.convert_from_old_format_wo_validation(content)
+    descr = InvalidDescr.model_validate(converted_content)
+
+    updated_content = dump_description(descr)
+    if output_path is not None:
+        write_yaml(updated_content, output_path)
+
+    return updated_content
 
 
 def ensure_description_is_model(