+++ tests/test_model/test_v0_5.py
@@ -233,7 +233,6 @@ def model():
             documentation=UNET2D_ROOT / "README.md",
             license=LicenseId("MIT"),
             git_repo=HttpUrl("https://github.com/bioimage-io/core-bioimage-io-python"),
-            format_version="0.5.3",
             description="description",
             authors=[
                 Author(name="Author 1", affiliation="Affiliation 1"),
@@ -567,3 +566,14 @@ def test_validate_parameterized_size(model: ModelDescr):
     param_size = model.inputs[0].axes[3].size
     assert isinstance(param_size, ParameterizedSize), type(param_size)
     assert (actual := param_size.validate_size(512)) == 512, actual
+
+
+def test_absolute_tolerance(model_data: Dict[str, Any]):
+    model_data["config"]["bioimageio"]["reproducibility_tolerance"] = [
+        {"absolute_tolerance": 100000}
+    ]
+    with ValidationContext(perform_io_checks=False):
+        model_descr = ModelDescr.model_validate(model_data)
+
+    with pytest.raises(ValueError), ValidationContext(perform_io_checks=True):
+        _ = model_descr._validate_test_tensors()  # type: ignore[reportPrivateUsage]